requirements.txt
PyQt6>=6.4.0
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0
keyboard>=0.13.5
pydirectinput>=1.0.4
mss>=9.0.1
Pillow>=10.0.0
